requirements.txt
fastapi[all]
stripe
psycopg2-binary
pytz
redmail
